opentelemetry_api
Module attributes

Available on crate feature metrics only.

OpenTelemetry Attributes

Structs

AttributeSet

An immutable set of distinct attributes.

DefaultAttributeEncoder

Default attribute encoding strategy

EncoderId

EncoderID is used to identify distinct Encoder implementations, for caching encoded results.

Iter

An iterator over the entries of a Set.

MergeIter

Merge two iterators, sorting by key

Traits

Encoder

Encoder is a mechanism for serializing an attribute set into a specific string representation that supports caching, to avoid repeated serialization. An example could be an exporter encoding the attribute set into a wire representation.

Functions

default_encoder

Build a new default encoder

hash_attributes

Impl of Hash for KeyValue

merge_iters

Merge two iterators, yielding sorted results

new_encoder_id

Build a new encoder id