opentelemetry_api::global
Trait ObjectSafeMeterProvider

pub trait ObjectSafeMeterProvider {
    // Required method
    fn versioned_meter_cow(
        &self,
        name: Cow<'static, str>,
        version: Option<Cow<'static, str>>,
        schema_url: Option<Cow<'static, str>>,
        attributes: Option<Vec<KeyValue>>,
    ) -> Meter;
}

Available on crate feature metrics only.

Allows a specific MeterProvider to be used generically by the GlobalMeterProvider by mirroring the interface and boxing the return types.

Required Methods

fn versioned_meter_cow( &self, name: Cow<'static, str>, version: Option<Cow<'static, str>>, schema_url: Option<Cow<'static, str>>, attributes: Option<Vec<KeyValue>>, ) -> Meter

Creates a versioned named meter instance that is a trait object through the underlying MeterProvider.