Enum opentelemetry_api::metrics::MetricsError

#[non_exhaustive]
pub enum MetricsError {
    Other(String),
    InvalidQuantile,
    NaNInput,
    NegativeInput,
    InconsistentAggregator(String),
    NoDataCollected,
    MetricKindMismatch(String),
    InconsistentState,
    NoSubtraction,
    ExportErr(Box<dyn ExportError>),
    InvalidInstrumentConfiguration(&'static str),
}

Available on crate feature metrics only.

Errors returned by the metrics API.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Other(String)

Other errors not covered by specific cases.

InvalidQuantile

Errors when requesting quantiles out of the 0-1 range.

NaNInput

Errors when recording nan values.

NegativeInput

Errors when recording negative values in monotonic sums.

InconsistentAggregator(String)

Errors when merging aggregators of incompatible types.

NoDataCollected

Errors when requesting data when no data has been collected

MetricKindMismatch(String)

Errors when registering to instruments with the same name and kind

InconsistentState

Errors when processor logic is incorrect

NoSubtraction

Errors when aggregator cannot subtract

ExportErr(Box<dyn ExportError>)

Fail to export metrics

InvalidInstrumentConfiguration(&'static str)

Invalid instrument configuration such invalid instrument name, invalid instrument description, invalid instrument unit, etc. See spec for full list of requirements.