Struct opentelemetry_api::metrics::InstrumentBuilder

pub struct InstrumentBuilder<'a, T> { /* private fields */ }

Available on crate feature metrics only.

Configuration for building an instrument.

Implementations

impl<'a, T> InstrumentBuilder<'a, T>
where T: TryFrom<Self, Error = MetricsError>,

pub fn with_description<S: Into<String>>(self, description: S) -> Self

Set the description for this instrument

pub fn with_unit(self, unit: Unit) -> Self

Set the unit for this instrument.

Unit is case sensitive(kb is not the same as kB).

Unit must be:

ASCII string

No longer than 63 characters

pub fn try_init(self) -> Result<T>

Validate the instrument configuration and creates a new instrument.

pub fn init(self) -> T

Creates a new instrument.

Panics

This function panics if the instrument configuration is invalid or the instrument cannot be created. Use try_init if you want to handle errors.