Module opentelemetry_api::trace::noop

Available on crate feature trace only.

No-op trace impls

This implementation is returned as the global tracer if no Tracer has been set. It is also useful for testing purposes as it is intended to have minimal resource utilization and runtime impact.

Structs

NoopSpan

A no-op instance of a Span.

NoopTextMapPropagator

A no-op instance of an TextMapPropagator.

NoopTracer

A no-op instance of a Tracer.

NoopTracerProvider

A no-op instance of a TracerProvider.