Struct opentelemetry_api::trace::OrderMap

pub struct OrderMap<K, V, S = RandomState>(_);

A hash table implementation that preserves insertion order across all operations.

Entries will be returned according to their insertion order when iterating over the collection.

Implementations

impl<K, V> OrderMap<K, V>

pub fn new() -> Self

Create a new map. (Does not allocate)

pub fn with_capacity(n: usize) -> Self

Create a new map with capacity for n key-value pairs. (Does not allocate if n is zero.)

Computes in O(n) time.

impl<K, V, S> OrderMap<K, V, S>

pub fn with_capacity_and_hasher(n: usize, hash_builder: S) -> Self

Create a new map with capacity for n key-value pairs. (Does not allocate if n is zero.)

Computes in O(n) time.

pub const fn with_hasher(hash_builder: S) -> Self

Create a new map with hash_builder.

This function is const, so it can be called in static contexts.

pub fn capacity(&self) -> usize

Computes in O(1) time.

pub fn hasher(&self) -> &S

Return a reference to the map’s BuildHasher.

pub fn len(&self) -> usize

Return the number of key-value pairs in the map.

Computes in O(1) time.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Computes in O(1) time.

pub fn iter(&self) -> Iter<'_, K, V>

Return an iterator over the key-value pairs of the map, in their order

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Return an iterator over the key-value pairs of the map, in their order

pub fn keys(&self) -> Keys<'_, K, V>

Return an iterator over the keys of the map, in their order

pub fn into_keys(self) -> IntoKeys<K, V>

Return an owning iterator over the keys of the map, in their order

pub fn values(&self) -> Values<'_, K, V>

Return an iterator over the values of the map, in their order

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>

Return an iterator over mutable references to the values of the map, in their order

pub fn into_values(self) -> IntoValues<K, V>

Return an owning iterator over the values of the map, in their order

pub fn clear(&mut self)

Remove all key-value pairs in the map, while preserving its capacity.

Computes in O(n) time.

pub fn truncate(&mut self, len: usize)

Shortens the map, keeping the first len elements and dropping the rest.

If len is greater than the map’s current length, this has no effect.

pub fn drain<R>(&mut self, range: R) -> Drain<'_, K, V> where
R: RangeBounds<usize>,

Clears the IndexMap in the given index range, returning those key-value pairs as a drain iterator.

The range may be any type that implements RangeBounds<usize>, including all of the std::ops::Range* types, or even a tuple pair of Bound start and end values. To drain the map entirely, use RangeFull like map.drain(..).

This shifts down all entries following the drained range to fill the gap, and keeps the allocated memory for reuse.

Panics if the starting point is greater than the end point or if the end point is greater than the length of the map.

pub fn split_off(&mut self, at: usize) -> Self where
S: Clone,

Splits the collection into two at the given index.

Returns a newly allocated map containing the elements in the range [at, len). After the call, the original map will be left containing the elements [0, at) with its previous capacity unchanged.

Panics if at > len.

impl<K, V, S> OrderMap<K, V, S> where
K: Hash + Eq,
S: BuildHasher,

pub fn reserve(&mut self, additional: usize)

Reserve capacity for additional more key-value pairs.

Computes in O(n) time.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the map as much as possible.

Computes in O(n) time.

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Insert a key-value pair in the map.

If an equivalent key already exists in the map: the key remains and retains in its place in the order, its corresponding value is updated with value and the older value is returned inside Some(_).

If no equivalent key existed in the map: the new key-value pair is inserted, last in order, and None is returned.

Computes in O(1) time (amortized average).

See also entry if you you want to insert or modify or if you need to get the index of the corresponding key-value pair.

pub fn insert_full(&mut self, key: K, value: V) -> (usize, Option<V>)

Insert a key-value pair in the map, and get their index.

If an equivalent key already exists in the map: the key remains and retains in its place in the order, its corresponding value is updated with value and the older value is returned inside (index, Some(_)).

If no equivalent key existed in the map: the new key-value pair is inserted, last in order, and (index, None) is returned.

Computes in O(1) time (amortized average).

See also entry if you you want to insert or modify or if you need to get the index of the corresponding key-value pair.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>

Get the given key’s corresponding entry in the map for insertion and/or in-place manipulation.

Computes in O(1) time (amortized average).

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
Q: Hash + Equivalent<K>,

Return true if an equivalent to key exists in the map.

Computes in O(1) time (average).

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
Q: Hash + Equivalent<K>,

Return a reference to the value stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_key_value<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where
Q: Hash + Equivalent<K>,

Return references to the key-value pair stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_full<Q: ?Sized>(&self, key: &Q) -> Option<(usize, &K, &V)> where
Q: Hash + Equivalent<K>,

Return item index, key and value

pub fn get_index_of<Q: ?Sized>(&self, key: &Q) -> Option<usize> where
Q: Hash + Equivalent<K>,

Return item index, if it exists in the map

Computes in O(1) time (average).

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where
Q: Hash + Equivalent<K>,

Return a mutable reference to the element pointed at by key, if it exists.

pub fn get_full_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(usize, &K, &mut V)> where
Q: Hash + Equivalent<K>,

Return a mutable reference to the element pointed at by key, if it exists. It also returns the element’s index and its key.

pub fn shift_remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V> where
Q: Hash + Equivalent<K>,

Remove the key-value pair equivalent to key and return its value.

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Return None if key is not in map.

Computes in O(n) time (average).

pub fn shift_remove_entry<Q: ?Sized>(&mut self, key: &Q) -> Option<(K, V)> where
Q: Hash + Equivalent<K>,

Remove and return the key-value pair equivalent to key.

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Return None if key is not in map.

Computes in O(n) time (average).

pub fn shift_remove_full<Q: ?Sized>(&mut self, key: &Q) -> Option<(usize, K, V)> where
Q: Hash + Equivalent<K>,

Remove the key-value pair equivalent to key and return it and the index it had.

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Return None if key is not in map.

Computes in O(n) time (average).

pub fn pop(&mut self) -> Option<(K, V)>

Remove the last key-value pair

This preserves the order of the remaining elements.

Computes in O(1) time (average).

pub fn retain<F>(&mut self, keep: F) where
F: FnMut(&K, &mut V) -> bool,

Scan through each key-value pair in the map and keep those where the closure keep returns true.

The elements are visited in order, and remaining elements keep their order.

Computes in O(n) time (average).

impl<K, V, S> OrderMap<K, V, S>

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Get a key-value pair by index

Valid indices are 0 <= index < self.len()

Computes in O(1) time.

pub fn get_index_mut(&mut self, index: usize) -> Option<(&mut K, &mut V)>

Get a key-value pair by index

Valid indices are 0 <= index < self.len()

Computes in O(1) time.

pub fn first(&self) -> Option<(&K, &V)>

Get the first key-value pair

Computes in O(1) time.

pub fn first_mut(&mut self) -> Option<(&K, &mut V)>

Get the first key-value pair, with mutable access to the value

Computes in O(1) time.

pub fn last(&self) -> Option<(&K, &V)>

Get the last key-value pair

Computes in O(1) time.

pub fn last_mut(&mut self) -> Option<(&K, &mut V)>

Get the last key-value pair, with mutable access to the value

Computes in O(1) time.

pub fn shift_remove_index(&mut self, index: usize) -> Option<(K, V)>

Remove the key-value pair by index

Valid indices are 0 <= index < self.len()

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Computes in O(n) time (average).