Crate opentelemetry_api

OpenTelemetry provides a single set of APIs, libraries, agents, and collector services to capture distributed traces and metrics from your application. You can analyze them using Prometheus, Jaeger, and other observability tools.

Compiler support: requires rustc 1.56+

Supported Rust Versions

OpenTelemetry is built against the latest stable release. The minimum supported version is 1.56. The current OpenTelemetry version is not guaranteed to build on Rust versions earlier than the minimum supported version.

The current stable Rust compiler and the three most recent minor versions before it will always be supported. For example, if the current stable compiler version is 1.49, the minimum supported version will not be increased past 1.46, three minor versions prior. Increasing the minimum supported compiler version is not considered a semver breaking change as long as doing so complies with this policy.

Modules

attributesmetrics

OpenTelemetry Attributes

baggage

Primitives for sending name/value data across system boundaries.

global

Utilities for working with global telemetry primitives

metricsmetrics

OpenTelemetry Metrics API

propagation

OpenTelemetry Propagator interface

tracetrace

API for tracing applications and libraries.

Structs

Context

An execution-scoped collection of values.

ContextGuard

A guard that resets the current context to the prior context when dropped.

InstrumentationLibrary

Information about a library or crate providing instrumentation.

Key

The key part of attribute KeyValue pairs.

KeyValue

A key-value pair describing an attribute.

StringValue

Wrapper for string-like values

Enums

Array

A Value::Array containing homogeneous values.

Value

The value part of attribute KeyValue pairs.

Traits

ExportError

Marker trait for errors returned by exporters