Module opentelemetry_api::metrics

OpenTelemetry Metrics API

Modules

noop

No-op OpenTelemetry Metrics Implementation

Structs

AsyncInstrumentBuilder

Configuration for building an async instrument.

Counter

An instrument that records increasing values.

Histogram

An instrument that records a distribution of values.

InstrumentBuilder

Configuration for building a sync instrument.

Meter

Provides access to instrument instances for recording metrics.

ObservableCounter

An async instrument that records increasing values.

ObservableGauge

An instrument that records independent readings.

ObservableUpDownCounter

An async instrument that records increasing or decreasing values.

Unit

Units denote underlying data units tracked by Meters.

UpDownCounter

An instrument that records increasing or decreasing values.

Enums

MetricsError

Errors returned by the metrics API.

Traits

AsyncInstrument

An SDK implemented instrument that records measurements via callback.

CallbackRegistration

A token representing the unique registration of a callback for a set of instruments with a Meter.

InstrumentProvider

SDK implemented trait for creating instruments

MeterProvider

Provides access to named Meter instances, for instrumenting an application or crate.

Observer

Records measurements for multiple instruments in a callback.

SyncCounter

An SDK implemented instrument that records increasing values.

SyncHistogram

An SDK implemented instrument that records a distribution of values.

SyncUpDownCounter

An SDK implemented instrument that records increasing or decreasing values.

Type Definitions

Callback

A function registered with a Meter that makes observations for the instruments it is registered with.

Result

A specialized Result type for metric operations.