Trait opentelemetry_api::metrics::MeterProvider

pub trait MeterProvider {
    // Required method
    fn versioned_meter(
        &self,
        name: &'static str,
        version: Option<&'static str>,
        schema_url: Option<&'static str>
    ) -> Meter;

    // Provided method
    fn meter(&self, name: &'static str) -> Meter { ... }
}

Available on crate feature metrics only.

Returns named meter instances

Required Methods

fn versioned_meter( &self, name: &'static str, version: Option<&'static str>, schema_url: Option<&'static str> ) -> Meter

Creates an implementation of the Meter interface.

The instrumentation name must be the name of the library providing instrumentation. This name may be the same as the instrumented code only if that code provides built-in instrumentation. If the instrumentation name is empty, then a implementation defined default name will be used instead.

Provided Methods

fn meter(&self, name: &'static str) -> Meter

Creates a named Meter instance.